+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## next
+
+- Fixed `parseChunked()` crash when input has trailing whitespaces
+
 ## 0.5.0 (2020-12-05)
 
 - Added support for Node.js 10